remove per frame deflate extension
