+++ jetty-websocket/websocket-core/src/main/java/org/eclipse/jetty/websocket/core/internal/WebSocketCoreSession.java
@@ -51,7 +51,6 @@ import org.eclipse.jetty.websocket.core.WebSocketConstants;
 import org.eclipse.jetty.websocket.core.WebSocketTimeoutException;
 import org.eclipse.jetty.websocket.core.WebSocketWriteTimeoutException;
 import org.eclipse.jetty.websocket.core.internal.Parser.ParsedFrame;
-import org.eclipse.jetty.websocket.core.internal.compress.DeflateFrameExtension;
 
 import static org.eclipse.jetty.util.Callback.NOOP;
 
@@ -589,9 +588,6 @@ public class WebSocketCoreSession implements IncomingFrames, FrameHandler.CoreSe
     @Override
     public void setAutoFragment(boolean autoFragment)
     {
-        // TODO: consider adding extensible/generic mechanism for extensions to validate configuration changes if more examples occur
-        if (autoFragment && getExtensionStack().getRsv1User() instanceof DeflateFrameExtension)
-            LOG.warn("Frame auto-fragmentation must not be used with DeflateFrameExtension");
         this.autoFragment = autoFragment;
     }
