--- jetty-websocket/websocket-core/src/main/java/org/eclipse/jetty/websocket/core/internal/compress/XWebkitDeflateFrameExtension.java
@@ -1,32 +1,0 @@
-//
-//  ========================================================================
-//  Copyright (c) 1995-2019 Mort Bay Consulting Pty. Ltd.
-//  ------------------------------------------------------------------------
-//  All rights reserved. This program and the accompanying materials
-//  are made available under the terms of the Eclipse Public License v1.0
-//  and Apache License v2.0 which accompanies this distribution.
-//
-//      The Eclipse Public License is available at
-//      http://www.eclipse.org/legal/epl-v10.html
-//
-//      The Apache License v2.0 is available at
-//      http://www.opensource.org/licenses/apache2.0.php
-//
-//  You may elect to redistribute this code under either of these licenses.
-//  ========================================================================
-//
-
-package org.eclipse.jetty.websocket.core.internal.compress;
-
-/**
- * Implementation of the <a href="https://tools.ietf.org/id/draft-tyoshino-hybi-websocket-perframe-deflate-05.txt">x-webkit-deflate-frame</a> extension seen out
- * in the wild. Using the alternate extension identification
- */
-public class XWebkitDeflateFrameExtension extends DeflateFrameExtension
-{
-    @Override
-    public String getName()
-    {
-        return "x-webkit-deflate-frame";
-    }
-}
