+++ jetty-websocket/websocket-core/src/main/resources/META-INF/services/org.eclipse.jetty.websocket.core.Extension
@@ -2,5 +2,3 @@ org.eclipse.jetty.websocket.core.internal.IdentityExtension
 org.eclipse.jetty.websocket.core.internal.FragmentExtension
 org.eclipse.jetty.websocket.core.internal.ValidationExtension
 org.eclipse.jetty.websocket.core.internal.compress.PerMessageDeflateExtension
-org.eclipse.jetty.websocket.core.internal.compress.DeflateFrameExtension
-org.eclipse.jetty.websocket.core.internal.compress.XWebkitDeflateFrameExtension
